changed to bigint

+++ src/datamodels/account_settings.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Column, ForeignKey
+from sqlalchemy import BIGINT, Column, ForeignKey
 
 from datamodels import Base, Jsonizable
 from datamodels.genshin_user import GenshinUser
@@ -8,6 +8,6 @@ class AccountInfo(Base):
     __tablename__ = "accountinfo"
 
     id = Column(
-        Integer, ForeignKey(GenshinUser.mihoyo_id, ondelete="CASCADE"), primary_key=True
+        BIGINT, ForeignKey(GenshinUser.mihoyo_id, ondelete="CASCADE"), primary_key=True
     )
     settings = Column(Jsonizable)
