+++ src/datamodels/birthday.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Column, Text, DateTime
+from sqlalchemy import BIGINT, Column, Text, DateTime
 
 from datamodels import Base
 
@@ -6,9 +6,9 @@ from datamodels import Base
 class Birthday(Base):
     __tablename__ = "bday"
 
-    discord_id = Column(Integer, primary_key=True)
-    guild_id = Column(Integer, primary_key=True)
-    month = Column(Integer, nullable=False)
-    day = Column(Integer, nullable=False)
+    discord_id = Column(BIGINT, primary_key=True)
+    guild_id = Column(BIGINT, primary_key=True)
+    month = Column(BIGINT, nullable=False)
+    day = Column(BIGINT, nullable=False)
     timezone = Column(Text, nullable=False)
     reminded_at = Column(DateTime)  # UTC timezone
