+++ src/datamodels/guild_settings.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Column, Text, String
+from sqlalchemy import BIGINT, Column, Text, String
 
 from datamodels import Base
 
@@ -6,7 +6,7 @@ from datamodels import Base
 class GuildSettings(Base):
     __tablename__ = "guildsettings"
 
-    guild_id = Column(Integer, primary_key=True)
+    guild_id = Column(BIGINT, primary_key=True)
     key = Column(String(100), primary_key=True)
     value = Column(Text)
