+++ src/datamodels/spiral_abyss.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Column, DateTime, Identity
+from sqlalchemy import BIGINT, Column, DateTime, Identity
 
 from datamodels import Base, Jsonizable
 
@@ -6,7 +6,7 @@ from datamodels import Base, Jsonizable
 class SpiralAbyssRotation(Base):
     __tablename__ = "abyssrotation"
 
-    id = Column(Integer, Identity(), primary_key=True)
+    id = Column(BIGINT, Identity(), primary_key=True)
     start = Column(DateTime, nullable=False, index=True)
     end = Column(DateTime, nullable=False, index=True)
     data = Column(Jsonizable)
