+++ src/datamodels/uid_mapping.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Column, Boolean, ForeignKey
+from sqlalchemy import BIGINT, Column, Boolean, ForeignKey
 
 from datamodels import Base
 
@@ -6,8 +6,8 @@ from datamodels import Base
 class UidMapping(Base):
     __tablename__ = "uidmapping"
 
-    uid = Column(Integer, primary_key=True)
+    uid = Column(BIGINT, primary_key=True)
     mihoyo_id = Column(
-        Integer, ForeignKey("genshinuser.mihoyo_id", ondelete="CASCADE"), nullable=False
+        BIGINT, ForeignKey("genshinuser.mihoyo_id", ondelete="CASCADE"), nullable=False
     )
     main = Column(Boolean, nullable=False, default=False)
